requirements.txt
gitdb==4.0.11
GitPython==3.1.40
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.16.1
requests==2.26.0
rich==13.6.0
smmap==5.0.1
